feat: #73 Extracting Skolemizer to separate module
